Add functionality to read Hibernate configuration
Now you can add "hibernate.properties" in the current working
directory, and it will apply all properties found in it, overwriting
properties in ".../resources/hibernate.properties". If it's not found,
then it doesn't do anything. The default properties file is always loaded
first.

When running the server code (without JAR), it will look for the file
in "server/". When running a JAR it's looking in the current working
directory.