Add show task functionality in manager's page
Also changes the way tasks and requests are recieved from the server.
With the tasks string sent in {message.data} it also sends the IDs of
the tasks in {message.lst}.

The show task button (only in manager's page) sends a request with the
specific id/num of the task instead of the index in the client. Because
requests can appear differently on every manager's client. Plus, we can
use the task id to Accept/Reject later.